rename spary and refactored utils
